refactor(dbeaver): change versioning system from tag to folder
Create subfolders in /dbeaver for the 7.2.4 and 7.2.5 versions. Each
subfolder contains a Dockerfile configured for the respective version.

The images, shortcut and dbeaver script continue in the /dbeaver folder.
It will be reused by all versions.

It too do a small change in the dbeaver script and in the "Create a
shortcut on the desktop" section. Now the DBEAVER_VERSION variable value will
be replaced during the creation of the shortcut.